+++ src/main/java/org/rundeck/client/api/model/ApiToken.java
@@ -18,14 +18,23 @@ package org.rundeck.client.api.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ApiToken {
     private String id;
+    private String token;
     private String user;
+    private String creator;
+    private List<String> roles;
+    private Boolean expired;
+    private DateInfo expiration;
 
     @Override
     public String toString() {
-        return "API Token: " + getTruncatedId();
+        return "API Token: " + getTruncatedIdOrToken();
     }
     public String toFullString() {
         return "API Token: " + id;
@@ -36,8 +45,16 @@ public class ApiToken {
         return id != null ? id.substring(0, 5) + "*****" : null;
     }
 
+    public String getTruncatedIdOrToken() {
+        return id != null ? id.substring(0, 5) + "*****" : null != token ? token.substring(0, 5) + "*****" : "?";
+    }
+
     public String getId() {
         return id;
+    }
+
+    public String getIdOrToken() {
+        return id != null ? id : token != null ? token : "?";
     }
 
     public void setId(String id) {
@@ -50,5 +67,69 @@ public class ApiToken {
 
     public void setUser(String user) {
         this.user = user;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public Map toMap() {
+        HashMap<String, Object> map = new HashMap<>();
+        map.put("user", getUser());
+        if (null != getRoles() && !getRoles().isEmpty()) {
+            map.put("roles", getRoles());
+        }
+        if (null != getId()) {
+            map.put("id", getId());
+        }
+        if (null != getToken()) {
+            map.put("token", getToken());
+        }
+        if (null != getCreator()) {
+            map.put("creator", getCreator());
+        }
+        if (null != getExpiration()) {
+            map.put("expiration", getExpiration().date);
+        }
+        if (null != getExpired()) {
+            map.put("expired", getExpired());
+        }
+        return map;
+    }
+
+    public String getCreator() {
+        return creator;
+    }
+
+    public void setCreator(String creator) {
+        this.creator = creator;
+    }
+
+    public List<String> getRoles() {
+        return roles;
+    }
+
+    public void setRoles(List<String> roles) {
+        this.roles = roles;
+    }
+
+    public Boolean getExpired() {
+        return expired;
+    }
+
+    public void setExpired(Boolean expired) {
+        this.expired = expired;
+    }
+
+    public DateInfo getExpiration() {
+        return expiration;
+    }
+
+    public void setExpiration(DateInfo expiration) {
+        this.expiration = expiration;
     }
 }