+++ src/main/java/org/rundeck/client/api/model/CreateToken.java
@@ -1,0 +1,20 @@
+package org.rundeck.client.api.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Value;
+
+import java.util.List;
+
+/**
+ * @author greg
+ * @since 5/3/17
+ */
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Value
+public class CreateToken {
+    String user;
+    List<String> roles;
+    String duration;
+}